split Service: one servce - one class ( and file) Create SessionClient in RimWorldOnlineCity remove obsolete method
