+++ Source/Converter/Converter.csproj
@@ -50,7 +50,7 @@
     <None Include="App.config" />
   </ItemGroup>
   <ItemGroup>
-    <ProjectReference Include="..\RimWorldOnlineCity\ServerCore\ServerCore.csproj">
+    <ProjectReference Include="..\ServerCore\ServerCore.csproj">
       <Project>{edc6d479-883b-4087-b9e1-47383b6713ee}</Project>
       <Name>ServerCore</Name>
     </ProjectReference>
