+++ Source/ServerCore/Common/StaticHelper.cs
@@ -1,0 +1,21 @@
+﻿using ServerOnlineCity.Model;
+using System.Collections.Generic;
+using System.Linq;
+
+namespace ServerOnlineCity.Common
+{
+    internal class StaticHelper
+    {
+        /// <summary>
+        /// Ники тех кого мы видим
+        /// </summary>
+        /// <returns></returns>
+        public static List<string> PartyLoginSee(PlayerServer player)
+        {
+            var ps = player.IsAdmin
+                ? Repository.GetData.PlayersAll.Select(p => p.Public.Login).ToList()
+                : player.Chats[0].PartyLogin;
+            return ps;
+        }
+    }
+}
