+++ Source/ServerCore/Model/ServerSettings.cs
@@ -1,9 +1,11 @@
-﻿namespace ServerCore.Model
+﻿using Transfer;
+
+namespace ServerCore.Model
 {
     public class ServerSettings
     {
         public string ServerName = "Another OnlineCity Server";
         public int SaveInterval = 10000;
-        public int Port = 8888;
+        public int Port = SessionClient.DefaultPort;
     }
 }
