+++ Source/ServerCore/Properties/launchSettings.json
@@ -1,7 +1,8 @@
 {
   "profiles": {
     "ServerCore": {
-      "commandName": "Project"
+      "commandName": "Project",
+      "commandLineArgs": "C:\\Word"
     }
   }
 }
