+++ Source/ServerCore/Services/GetPlayerByToken.cs
@@ -1,0 +1,30 @@
+﻿using System;
+using System.Linq;
+using OCUnion.Transfer.Model;
+using ServerOnlineCity.Model;
+using Transfer;
+
+namespace ServerOnlineCity.Services
+{
+    internal sealed class GetPlayerByToken : IGenerateResponseContainer
+    {
+        public int RequestTypePackage => (int)PackageType.RequestPlayerByToken;
+
+        public int ResponseTypePackage => (int)PackageType.ResponsePlayerByToken;
+
+        public ModelContainer GenerateModelContainer(ModelContainer modelContainer, ref PlayerServer player)
+        {
+            if (player == null) return null;
+            var packet = (Guid)modelContainer.Packet;
+            var result = new ModelContainer { TypePacket = ResponseTypePackage };
+
+            lock (player)
+            {
+                var data = Repository.GetData;
+                var playerServer = data.PlayersAll.FirstOrDefault(p => packet.Equals(p.DiscordToken));
+                result.Packet = playerServer?.Public;
+                return result;
+            }
+        }
+    }
+}
