+++ Source/ServerCore/Services/IGenerateResponseContainer.cs
@@ -1,0 +1,12 @@
+﻿using ServerOnlineCity.Model;
+using Transfer;
+
+namespace ServerOnlineCity.Services
+{
+    public interface IGenerateResponseContainer
+    {
+        int RequestTypePackage { get; }
+        int ResponseTypePackage { get; }
+        ModelContainer GenerateModelContainer(ModelContainer request, ref PlayerServer player);
+    }
+}
