+++ Source/ServerCore/Services/PlayInfo.cs
@@ -1,0 +1,171 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using Model;
+using OCUnion;
+using ServerOnlineCity.Common;
+using ServerOnlineCity.Model;
+using Transfer;
+
+namespace ServerOnlineCity.Services
+{
+    internal sealed class PlayInfo : IGenerateResponseContainer
+    {
+        public int RequestTypePackage => 11;
+
+        public int ResponseTypePackage => 12;
+
+        public ModelContainer GenerateModelContainer(ModelContainer request, ref PlayerServer player)
+        {
+            if (player == null) return null;
+            var result = new ModelContainer() { TypePacket = ResponseTypePackage };
+            result.Packet = playInfo((ModelPlayToServer)request.Packet, ref player);
+            return result;
+        }
+
+        public ModelPlayToClient playInfo(ModelPlayToServer packet, ref PlayerServer player)
+        {
+            lock (player)
+            {
+                var data = Repository.GetData;
+                var timeNow = DateTime.UtcNow;
+                var toClient = new ModelPlayToClient();
+                toClient.UpdateTime = timeNow;
+                if (packet.GetPlayersInfo != null && packet.GetPlayersInfo.Count > 0)
+                {
+                    var pSee = StaticHelper.PartyLoginSee(player);
+                    var pGet = pSee.Where(s => packet.GetPlayersInfo.Any(p => s == p)).ToList();
+                    toClient.PlayersInfo = Repository.GetData.PlayersAll
+                        .Where(p => pGet.Any(g => g == p.Public.Login))
+                        .Select(p => p.Public)
+                        .ToList();
+                    if (pGet.Any(p => p == null)) Loger.Log("nulllllll");
+                }
+                if (packet.SaveFileData != null && packet.SaveFileData.Length > 0)
+                {
+                    player.SaveDataPacket = packet.SaveFileData;
+                    player.Public.LastSaveTime = timeNow;
+                    Repository.Get.ChangeData = true;
+                }
+                player.Public.LastTick = packet.LastTick;
+
+                var pLogin = player.Public.Login;
+                //packet.WObjects тут все объекты этого игрока, добавляем которых у нас нет
+                var pWOs = packet.WObjects ?? new List<WorldObjectEntry>();
+                //packet.WObjectsToDelete тут те объекты этого игрока, что нужно удалить
+                var pDs = packet.WObjectsToDelete ?? new List<WorldObjectEntry>();
+                //передаем назад объекты у которых не было ServerId, т.е. они новые для сервера + все с изменениями
+                var outWO = new List<WorldObjectEntry>();
+                var outWOD = new List<WorldObjectEntry>();
+                //это первое обращение, если не прислали своих объектов
+                var first = pWOs.Count == 0;
+                lock (data)
+                {
+                    for (int i = 0; i < pDs.Count; i++)
+                    {
+                        if (pDs[i].LoginOwner != player.Public.Login) continue;
+                        var sid = pDs[i].ServerId;
+                        var pD = data.WorldObjects.FirstOrDefault(p => p.ServerId == sid);
+                        if (pD != null)
+                        {
+                            //удаление из базы
+                            pD.UpdateTime = timeNow;
+                            data.WorldObjects.Remove(pD);
+                            data.WorldObjectsDeleted.Add(pD);
+                        }
+                    }
+
+                    for (int i = 0; i < pWOs.Count; i++)
+                    {
+                        if (pWOs[i].LoginOwner != player.Public.Login) continue; // <-на всякий случай
+                        var sid = pWOs[i].ServerId;
+                        if (sid == 0)
+                        {
+                            //добавление в базу
+                            pWOs[i].UpdateTime = timeNow;
+                            pWOs[i].ServerId = data.GetWorldObjectEntryId();
+                            data.WorldObjects.Add(pWOs[i]);
+                            outWO.Add(pWOs[i]);
+                            continue;
+                        }
+                        var WO = data.WorldObjects.FirstOrDefault(p => p.ServerId == sid);
+                        if (WO != null)
+                        {
+                            //данный объект уже есть в базу обновляем по нему информкацию
+                            if (WO.Name != pWOs[i].Name)
+                            {
+                                WO.UpdateTime = timeNow;
+                                WO.Name = pWOs[i].Name;
+                            }
+                            if (WO.FreeWeight != pWOs[i].FreeWeight)
+                            {
+                                WO.UpdateTime = timeNow;
+                                WO.FreeWeight = pWOs[i].FreeWeight;
+                            }
+                            if (WO.MarketValue != pWOs[i].MarketValue)
+                            {
+                                WO.UpdateTime = timeNow;
+                                WO.MarketValue = pWOs[i].MarketValue;
+                            }
+                            if (WO.MarketValuePawn != pWOs[i].MarketValuePawn)
+                            {
+                                WO.UpdateTime = timeNow;
+                                WO.MarketValuePawn = pWOs[i].MarketValuePawn;
+                            }
+                            if (WO.Tile != pWOs[i].Tile)
+                            {
+                                WO.UpdateTime = timeNow;
+                                WO.Tile = pWOs[i].Tile;
+                            }
+                        }
+                        else
+                        {
+                            Loger.Log("PlayInfo find error add WO: " + pWOs[i].Name + " sid=" + sid);
+                        }
+                    }
+
+                    //передаем объекты других игроков + при первом обращении first
+                    for (int i = 0; i < data.WorldObjects.Count; i++)
+                    {
+                        if (data.WorldObjects[i].UpdateTime < packet.UpdateTime) continue;
+                        if (!first && data.WorldObjects[i].LoginOwner == pLogin) continue;
+                        outWO.Add(data.WorldObjects[i]);
+                    }
+
+                    //передаем удаленные объекты других игроков (не для первого запроса)
+                    if (packet.UpdateTime > DateTime.MinValue)
+                    {
+                        for (int i = 0; i < data.WorldObjectsDeleted.Count; i++)
+                        {
+                            if (data.WorldObjectsDeleted[i].UpdateTime < packet.UpdateTime)
+                            {
+                                //Удаляем все записи сроком старше 2х минут (их нужно хранить время между тем как игрок у которого удалился караван зальёт это на сервер, и все другие онлайн игроки получат эту инфу, а обновление идет раз в 5 сек)
+                                if ((timeNow - data.WorldObjectsDeleted[i].UpdateTime).TotalSeconds > 120000)
+                                {
+                                    data.WorldObjectsDeleted.RemoveAt(i--);
+                                }
+                                continue;
+                            }
+                            if (data.WorldObjectsDeleted[i].LoginOwner == pLogin) continue;
+                            outWOD.Add(data.WorldObjectsDeleted[i]);
+                        }
+                    }
+
+                    //завершили сбор информации клиенту
+                    toClient.WObjects = outWO;
+                    toClient.WObjectsToDelete = outWOD;
+                    player.LastUpdateTime = timeNow;
+                }
+
+                //прикрепляем письма
+                toClient.Mails = player.Mails;
+                player.Mails = new List<ModelMailTrade>();
+
+                //флаг, что на клиента кто-то напал и он должен запросить подробности
+                toClient.AreAttacking = player.AttackData != null && player.AttackData.Host == player && player.AttackData.State == 1;
+
+                return toClient;
+            }
+        }
+    }
+}
